add initial generated services
