+++ pkg/services/kubernetes/v1.4/config.yaml
@@ -1,5 +1,5 @@
 package: kubernetes
-output: src/kubernetes.gen.go
+output: generated/kubernetes.gen.go
 generate:
   models: true
   client: true
